added cmake_helper.py
