+++ amigo/model.py
@@ -20,7 +20,7 @@ from .amigo import (
     CSRMat,
     ExternalComponentGroup,
 )
-from .utils import get_cmake_dir
+from .cmake_helper import get_cmake_dir
 from .component import Component
 
 try:
@@ -1077,9 +1077,6 @@ amigo_add_python_module(
 
         # Locate the installed Amigo CMake package inside the Python package
         amigo_cmake_dir = get_cmake_dir()
-
-        print(amigo_cmake_dir)
-        exit(0)
 
         # Configure
         p = subprocess.Popen(
